перенес код карты в отдельный модуль
